+++ src/libecalc/domain/infrastructure/energy_components/electricity_consumer/electricity_consumer.py
@@ -18,7 +18,7 @@ from libecalc.domain.infrastructure.energy_components.utils import (
     _convert_keys_in_dictionary_from_str_to_periods,
     check_model_energy_usage_type,
 )
-from libecalc.domain.process.dto import ElectricEnergyUsageModel
+from libecalc.domain.process.dto.energy_usage_model_types import ElectricEnergyUsageModel
 from libecalc.domain.process.process_system import ProcessSystem
 from libecalc.dto.types import ConsumerUserDefinedCategoryType
 from libecalc.dto.utils.validators import validate_temporal_model
