+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/direct_consumer_function.py
@@ -1,7 +1,7 @@
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function.direct_expression_consumer_function import (
     DirectExpressionConsumerFunction,
 )
-from libecalc.domain.process.dto import DirectConsumerFunction
+from libecalc.domain.process.dto.direct import DirectConsumerFunction
 
 
 def create_direct_consumer_function(model_dto: DirectConsumerFunction) -> DirectExpressionConsumerFunction:
