+++ src/libecalc/domain/process/compressor/core/train/variable_speed_compressor_train_common_shaft_multiple_streams_and_pressures.py
@@ -28,15 +28,15 @@ from libecalc.domain.process.compressor.core.train.utils.numeric_methods import 
 from libecalc.domain.process.compressor.core.train.utils.variable_speed_compressor_train_common_shaft import (
     get_single_speed_equivalent,
 )
+from libecalc.domain.process.compressor.dto.stage import CompressorStage
+from libecalc.domain.process.compressor.dto.train import (
+    SingleSpeedCompressorTrain,
+    VariableSpeedCompressorTrainMultipleStreamsAndPressures,
+)
 from libecalc.domain.process.core import ModelInputFailureStatus, validate_model_input
 from libecalc.domain.process.core.results import CompressorTrainResult
 from libecalc.domain.process.core.results.compressor import (
     TargetPressureStatus,
-)
-from libecalc.domain.process.dto import (
-    CompressorStage,
-    SingleSpeedCompressorTrain,
-    VariableSpeedCompressorTrainMultipleStreamsAndPressures,
 )
 
 EPSILON = 1e-5
