+++ src/libecalc/presentation/yaml/domain/reference_service.py
@@ -1,7 +1,8 @@
 from collections.abc import Iterable
 from typing import Protocol
 
-from libecalc.domain.process.dto import CompressorModelTypes, GeneratorSetSampled, TabulatedData
+from libecalc.domain.process.compressor.dto.compressor_models import CompressorModelTypes
+from libecalc.domain.process.dto import GeneratorSetSampled, TabulatedData
 from libecalc.domain.process.pump.pump import PumpModelDTO
 from libecalc.dto import FuelType
