+++ tests/libecalc/core/models/compressor_modelling/sampled/test_compressor_sampled.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 import libecalc.common.energy_usage_type
-from libecalc.domain.process import dto
+from libecalc.domain.process.compressor.dto import compressor_models as dto
 from libecalc.common.units import Unit
 from libecalc.domain.process.compressor.core.sampled import CompressorModelSampled
 from libecalc.domain.process.compressor.core.sampled.compressor_model_sampled_1d import (
