+++ tests/libecalc/core/models/compressor_modelling/test_compressor_train_base.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import numpy as np
 import pytest
 
-from libecalc.domain.process import dto
+from libecalc.domain.process.compressor.dto import compressor_models as dto
 from libecalc.domain.process.compressor.core.train.base import CompressorTrainModel
